# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "additionalProperties": false,
        "description": "クライアントが送る会話履歴の1件(仕様§6.3: サーバは履歴を保存しない)。",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "history": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "History",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "graphs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Provenance"
            },
            "title": "Graphs",
            "type": "object"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ChatSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "tool_call_limit": {
            "title": "Tool Call Limit",
            "type": "integer"
          },
          "tool_call_limit_reached": {
            "title": "Tool Call Limit Reached",
            "type": "boolean"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallLogEntry"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "answer",
          "sources",
          "graphs",
          "tool_calls",
          "tool_call_limit",
          "tool_call_limit_reached",
          "truncated"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatSource": {
        "additionalProperties": false,
        "description": "回答の根拠として実際に道具が返したエンティティ1件(裁定B92裁定2(2))。\n\n**LLMの発言からではなく、道具の戻り値から`chat_tools.SourceCollector`が\n機械的に組み立てる。** 道具が0件を返した実行では、この一覧は必ず空になる\n——空にならない実行があれば捏造(`tests/test_api_chat.py`の壊し確認)。",
        "properties": {
          "graphs": {
            "items": {
              "type": "string"
            },
            "title": "Graphs",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "id_path": {
            "title": "Id Path",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "id_path",
          "type",
          "label",
          "graphs"
        ],
        "title": "ChatSource",
        "type": "object"
      },
      "EntityDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "attributes": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/AttributeValue"
              },
              "type": "array"
            },
            "title": "Attributes",
            "type": "object"
          },
          "graphs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Provenance"
            },
            "title": "Graphs",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "id_path": {
            "title": "Id Path",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "relationships": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Relationship"
              },
              "type": "array"
            },
            "title": "Relationships",
            "type": "object"
          },
          "relationships_limit": {
            "title": "Relationships Limit",
            "type": "integer"
          },
          "relationships_truncated": {
            "title": "Relationships Truncated",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "id_path",
          "type",
          "label",
          "attributes",
          "relationships",
          "graphs",
          "relationships_limit",
          "relationships_truncated"
        ],
        "title": "EntityDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NeighborhoodResponse": {
        "additionalProperties": false,
        "description": "近傍サブグラフ(仕様§9.1「指定ノードから深さ1-2のノード/エッジ」)。\n\n**既知の限界(観察O10)**: `LawRevision`(実データで9,550件)は\n**グラフの辺を1本も持たない** —— 法令への結びつきは`lawId`という\nリテラルとIRIのパスに埋まった法令IDだけで、辺として存在しない。\n**したがって法令の近傍に改正は一度も現れない。**\n利用者が「この法令の近傍にはこれしか無い」と読むのは誤りである\n(辺を足すのはPhase 2の課題)。**黙っていないためにここに書く**\n(このプロジェクトの再発欠陥6: 報告が嘘をつく)。",
        "properties": {
          "center": {
            "$ref": "#/components/schemas/EntityRef"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "edge_limit": {
            "title": "Edge Limit",
            "type": "integer"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "edges_truncated": {
            "title": "Edges Truncated",
            "type": "boolean"
          },
          "fanout_limit": {
            "title": "Fanout Limit",
            "type": "integer"
          },
          "fanout_truncated_nodes": {
            "items": {
              "type": "string"
            },
            "title": "Fanout Truncated Nodes",
            "type": "array"
          },
          "graphs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Provenance"
            },
            "title": "Graphs",
            "type": "object"
          },
          "node_limit": {
            "title": "Node Limit",
            "type": "integer"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "title": "Nodes",
            "type": "array"
          },
          "nodes_truncated": {
            "title": "Nodes Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "center",
          "depth",
          "nodes",
          "edges",
          "graphs",
          "node_limit",
          "edge_limit",
          "fanout_limit",
          "nodes_truncated",
          "edges_truncated",
          "fanout_truncated_nodes"
        ],
        "title": "NeighborhoodResponse",
        "type": "object"
      },
      "PathResponse": {
        "additionalProperties": false,
        "description": "パス探索(仕様§9.1「2エンティティ間の経路(法令↔法人など)」)。\n\n**`found=False`の読み方を応答の形で強制する。**\n「max_depth以内に経路が無かった」と「経路が存在しない」は違う ——\n空の結果だけを返すと利用者は後者だと読む(このプロジェクトが繰り返し\n重い欠陥として扱ってきた「報告が嘘をつく」型)。`exhaustive`が真で\n初めて後者を意味する。",
        "properties": {
          "budget_exhausted": {
            "title": "Budget Exhausted",
            "type": "boolean"
          },
          "depth_limited": {
            "title": "Depth Limited",
            "type": "boolean"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "exhaustive": {
            "title": "Exhaustive",
            "type": "boolean"
          },
          "fanout_limit": {
            "title": "Fanout Limit",
            "type": "integer"
          },
          "fanout_truncated": {
            "title": "Fanout Truncated",
            "type": "boolean"
          },
          "found": {
            "title": "Found",
            "type": "boolean"
          },
          "goal": {
            "$ref": "#/components/schemas/EntityRef"
          },
          "graphs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Provenance"
            },
            "title": "Graphs",
            "type": "object"
          },
          "max_depth": {
            "title": "Max Depth",
            "type": "integer"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "title": "Nodes",
            "type": "array"
          },
          "searched_depth": {
            "title": "Searched Depth",
            "type": "integer"
          },
          "start": {
            "$ref": "#/components/schemas/EntityRef"
          },
          "undirected": {
            "title": "Undirected",
            "type": "boolean"
          },
          "visit_budget": {
            "title": "Visit Budget",
            "type": "integer"
          },
          "visited": {
            "title": "Visited",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "goal",
          "nodes",
          "edges",
          "graphs",
          "found",
          "max_depth",
          "visit_budget",
          "visited",
          "searched_depth",
          "budget_exhausted",
          "depth_limited",
          "fanout_limit",
          "fanout_truncated",
          "exhaustive",
          "undirected"
        ],
        "title": "PathResponse",
        "type": "object"
      },
      "Provenance": {
        "additionalProperties": false,
        "description": "CQ7(`queries/cq/cq07-provenance-of-edge.rq`)と同じ3項+グラフ自身。",
        "properties": {
          "available": {
            "default": true,
            "title": "Available",
            "type": "boolean"
          },
          "fetched_on": {
            "title": "Fetched On",
            "type": "string"
          },
          "graph": {
            "title": "Graph",
            "type": "string"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "graph",
          "source",
          "fetched_on",
          "license"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Results",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "query",
          "results",
          "limit",
          "truncated"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "ToolCallLogEntry": {
        "additionalProperties": false,
        "description": "道具呼び出し1件の監査記録(裁定B92裁定2(3): 調査の過程を監査できる)。",
        "properties": {
          "arguments": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Arguments",
            "type": "object"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool",
          "arguments",
          "result_count"
        ],
        "title": "ToolCallLogEntry",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Japan Government KG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat": {
      "post": {
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat"
      }
    },
    "/entity/{entity_id}": {
      "get": {
        "operationId": "entity_detail_entity__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "関係一覧の既定50・最大200。超過は422で拒否する",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "関係一覧の既定50・最大200。超過は422で拒否する",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity Detail"
      }
    },
    "/neighborhood/{entity_id}": {
      "get": {
        "operationId": "neighborhood_neighborhood__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "深さ。既定1・最大2(仕様§9.1が「深さ1-2」と定めている)。超過は422で拒否する(黙って丸めない)",
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "description": "深さ。既定1・最大2(仕様§9.1が「深さ1-2」と定めている)。超過は422で拒否する(黙って丸めない)",
              "maximum": 2,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "description": "総ノード数の上限。既定100・最大500",
            "in": "query",
            "name": "node_limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "総ノード数の上限。既定100・最大500",
              "maximum": 500,
              "minimum": 1,
              "title": "Node Limit",
              "type": "integer"
            }
          },
          {
            "description": "総エッジ数の上限。既定200・最大1000",
            "in": "query",
            "name": "edge_limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "総エッジ数の上限。既定200・最大1000",
              "maximum": 1000,
              "minimum": 1,
              "title": "Edge Limit",
              "type": "integer"
            }
          },
          {
            "description": "**1ノードあたりの分岐数の上限。** 総数の上限だけではハブ1個が予算を食い潰し、他の方向が1つも見えなくなる(既定25・最大100)",
            "in": "query",
            "name": "fanout_limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "**1ノードあたりの分岐数の上限。** 総数の上限だけではハブ1個が予算を食い潰し、他の方向が1つも見えなくなる(既定25・最大100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Fanout Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Neighborhood"
      }
    },
    "/path": {
      "get": {
        "operationId": "path_path_get",
        "parameters": [
          {
            "description": "始点の`id_path`(`/entity/`と同じ形)",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "description": "始点の`id_path`(`/entity/`と同じ形)",
              "minLength": 1,
              "title": "From",
              "type": "string"
            }
          },
          {
            "description": "終点の`id_path`(`/entity/`と同じ形)",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "description": "終点の`id_path`(`/entity/`と同じ形)",
              "minLength": 1,
              "title": "To",
              "type": "string"
            }
          },
          {
            "description": "探索の深さ。既定4(法令→府省→事業→支出→法人の縦スライスが4ホップ)・最大6。超過は422で拒否する",
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "default": 4,
              "description": "探索の深さ。既定4(法令→府省→事業→支出→法人の縦スライスが4ホップ)・最大6。超過は422で拒否する",
              "maximum": 6,
              "minimum": 1,
              "title": "Max Depth",
              "type": "integer"
            }
          },
          {
            "description": "訪問ノード数の予算。既定400・最大2000。**使い切ったら`budget_exhausted`が真になり、`found=false`は「無い」ではなく「見つからなかった」を意味する**",
            "in": "query",
            "name": "visit_budget",
            "required": false,
            "schema": {
              "default": 400,
              "description": "訪問ノード数の予算。既定400・最大2000。**使い切ったら`budget_exhausted`が真になり、`found=false`は「無い」ではなく「見つからなかった」を意味する**",
              "maximum": 2000,
              "minimum": 2,
              "title": "Visit Budget",
              "type": "integer"
            }
          },
          {
            "description": "1ノードあたりの分岐数の上限。既定50・最大200。**これが効くと探索は不完全になり`exhaustive`は真になれない**",
            "in": "query",
            "name": "fanout_limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "1ノードあたりの分岐数の上限。既定50・最大200。**これが効くと探索は不完全になり`exhaustive`は真になれない**",
              "maximum": 200,
              "minimum": 1,
              "title": "Fanout Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Path"
      }
    },
    "/search": {
      "get": {
        "operationId": "search_search_get",
        "parameters": [
          {
            "description": "検索語(部分一致・大文字小文字を無視)",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "検索語(部分一致・大文字小文字を無視)",
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "既定20・最大100。超過は422で拒否する(黙って丸めない。queries.pyのコメント参照)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "既定20・最大100。超過は422で拒否する(黙って丸めない。queries.pyのコメント参照)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search"
      }
    }
  }
}